1.1.0文件提取功能可以提取data下的文件

+++ app/src/main/java/xyz/junerver/sfbrowser/BrowserItemOnClickOnClickListenerImpl.kt
@@ -4,10 +4,13 @@ import android.content.ClipData
 import android.content.ClipboardManager
 import android.content.Context
 import android.content.Intent
+import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import com.lxj.xpopup.XPopup
 import xyz.junerver.fileselector.*
+import java.io.BufferedInputStream
 import java.io.File
+import java.io.FileOutputStream
 import java.nio.file.Files
 
 /**
@@ -17,7 +20,9 @@ import java.nio.file.Files
  * Email: junerver@gmail.com
  * Version: v1.0
  */
-class BrowserItemOnClickOnClickListenerImpl(private val mCallBack: OperateFileModelItemCallBack) : FileAdapter.BrowserItemOnClickListener {
+class BrowserItemOnClickOnClickListenerImpl(private val mCallBack: OperateFileModelItemCallBack) :
+    FileAdapter.BrowserItemOnClickListener {
+
     override fun onItemClick(ctx: Context, fileModel: FileModel) {
         val ex = getExtension(fileModel.path).lowercase()
         if (FileSelector.IMAGE_TYPES.contains(ex)) {
@@ -26,43 +31,92 @@ class BrowserItemOnClickOnClickListenerImpl(private val mCallBack: OperateFileMo
             i.putExtra("path", fileModel.path)
             ctx.startActivity(i)
         } else {
-            ctx.openFile(fileModel.path)
+            if (!fileModel.isAndroidData) {
+                ctx.openFile(fileModel.path)
+            } else {
+                ctx.toast("无法打开Android/data目录下的文件！请长按使用提取功能!")
+            }
         }
     }
 
     override fun onItemLongClick(ctx: Context, fileModel: FileModel) {
-        //TODO("弹出操作对话框：提取、删除、复制路径、查看详情、重命名")
         XPopup.Builder(ctx)
             .asBottomList(
-                "", arrayOf("提取", "删除", "复制路径", "查看详情", "重命名")
-            ) { position, _ ->
+                "", MENU_ITEMS
+            ) { _, text ->
                 val oldFile = File(fileModel.path)
-                when (position) {
-                    0 -> {
+                when (text) {
+                    EXTRACT_FILE -> {
                         //复制文件到SF的目录下
                         val target = File(ROOT_DIR + fileModel.name)
-                        if (oldFile.exists() && oldFile.isFile) {
-                            Files.copy(oldFile.toPath(), target.toPath())
-                            ctx.toast("文件提取成功：${target.path}")
+                        if (target.exists() && target.isFile && target.length() > 0) {
+                            ctx.toast("目标文件已存在！")
+                            return@asBottomList
+                        }
+                        if (!fileModel.isAndroidData) {
+                            if (oldFile.exists() && oldFile.isFile) {
+                                Files.copy(oldFile.toPath(), target.toPath())
+                                ctx.toast("文件提取成功：${target.path}")
+                            } else {
+                                ctx.toast("无效文件，操作失败！")
+                            }
                         } else {
-                            ctx.toast("无效文件，操作失败！")
+                            //
+                            GlobalThreadPools.getInstance().execute {
+                                fileModel.documentFile?.let { doc ->
+                                    val fis = ctx.contentResolver.openInputStream(doc.uri)
+                                    fis?.let { `is` ->
+                                        val bis = BufferedInputStream(`is`)
+                                        val fos = FileOutputStream(target)
+                                        try {
+                                            val buffer = ByteArray(1024)
+                                            var len: Int
+                                            var total = 0
+                                            while (((bis.read(buffer)).also { len = it }) != -1) {
+                                                fos.write(buffer, 0, len)
+                                                total += len
+                                                //获取当前下载量
+                                                val progress: Double =
+                                                    total.toDouble() / fileModel.size
+                                                "下载量： $progress".log()
+                                            }
+                                            postUI {
+                                                ctx.toast("文件提取成功：${target.path}")
+                                            }
+                                        } catch (e: Exception) {
+                                            postUI {
+                                                ctx.toast("文件提取失败！")
+                                            }
+                                        } finally {
+                                            fos.close()
+                                            bis.close()
+                                            `is`.close()
+                                        }
+                                    }
+                                }
+                            }
                         }
                     }
-                    1 -> {
-                        if (oldFile.exists() && oldFile.isFile) {
+                    DELETE -> {
+                        fun showDelDialog(delAct: () -> Unit) {
                             AlertDialog.Builder(ctx)
                                 .setTitle("警告：危险操作不可逆！")
                                 .setNeutralButton("取消") { _, _ -> }
                                 .setPositiveButton("确定") { _, _ ->
-                                    oldFile.delete()
+                                    delAct.invoke()
                                     mCallBack.delItem(fileModel)
                                     ctx.toast("删除成功！")
                                 }.show()
+                        }
+                        if (!fileModel.isAndroidData && oldFile.exists() && oldFile.isFile) {
+                            showDelDialog { oldFile.delete() }
+                        } else if (fileModel.isAndroidData && fileModel.documentFile != null) {
+                            showDelDialog { fileModel.documentFile!!.delete() }
                         } else {
                             ctx.toast("无效文件，操作失败！")
                         }
                     }
-                    2 -> {
+                    COPY_PATH -> {
                         //获取剪贴板管理器：
                         val cm: ClipboardManager =
                             ctx.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
@@ -72,28 +126,36 @@ class BrowserItemOnClickOnClickListenerImpl(private val mCallBack: OperateFileMo
                         cm.setPrimaryClip(mClipData)
                         ctx.toast("已复制路径：${fileModel.path}")
                     }
-                    3 -> {
+                    SEE_DETAIL -> {
                         XPopup.Builder(ctx)
                             .asCustom(FileDetailsPopup(ctx, fileModel))
                             .show()
                     }
-                    4 -> {
-                        if (oldFile.exists() && oldFile.isFile) {
+                    RENAME -> {
+                        fun showRenamePopup(rename: (newName: String) -> Unit) {
                             XPopup.Builder(ctx).asInputConfirm(
                                 "重命名", "", getNameExcludeExtension(fileModel.path), null
                             ) { text ->
                                 val newName = text + ".${fileModel.extension}"
                                 if (newName.isValidFileName()) {
-                                    val target =
-                                        File(oldFile.parent + File.separator + newName)
-                                    oldFile.renameTo(target)
-                                    fileModel.update(target)
+                                    rename(newName)
                                     mCallBack.changeItem(fileModel)
                                 } else {
                                     ctx.toast("文件名非法！")
                                 }
-
                             }.show()
+                        }
+                        if (oldFile.exists() && oldFile.isFile && !fileModel.isAndroidData) {
+                            showRenamePopup { newName ->
+                                val target = File(oldFile.parent + File.separator + newName)
+                                oldFile.renameTo(target)
+                                fileModel.update(target)
+                            }
+                        } else if (fileModel.isAndroidData && fileModel.documentFile != null) {
+                            showRenamePopup {
+                                fileModel.documentFile!!.renameTo(it)
+                                fileModel.name = it
+                            }
                         } else {
                             ctx.toast("无效文件，操作失败！")
                         }
@@ -101,6 +163,15 @@ class BrowserItemOnClickOnClickListenerImpl(private val mCallBack: OperateFileMo
                 }
             }
             .show()
+    }
+
+    companion object {
+        private const val EXTRACT_FILE = "提取"
+        private const val DELETE = "删除"
+        private const val COPY_PATH = "复制路径"
+        private const val SEE_DETAIL = "查看详情"
+        private const val RENAME = "重命名"
+        private val MENU_ITEMS = arrayOf(EXTRACT_FILE, DELETE, COPY_PATH, SEE_DETAIL, RENAME)
     }
 }
